Split off new AuthDocs to allow easy dev on v2
