+++ src/containers/documentation/AuthorizationDocsLegacy.scss
@@ -1,0 +1,47 @@
+@import '../../styles/variables';
+
+.va-api-authorization-docs-legacy {
+  p,
+  td {
+    line-height: 1.6;
+  }
+
+  code,
+  .code-wrapper {
+    // Needs both due to covering of both code blocks and inline formatted code/parameter names
+    background-color: #222935;
+    border-radius: 3px;
+  }
+
+  code {
+    color: $color-white;
+    font-family: 'Source Code Pro', monospace;
+    padding: 0.2rem 0.4rem;
+    margin: auto 0.2rem;
+
+    // this may look ugly but there's no padding at line end/start when text wraps
+    white-space: nowrap;
+  }
+
+  pre > code {
+    display: block;
+    padding: 2rem;
+    margin: 0;
+    white-space: pre;
+
+    &:focus {
+      border: 1px solid $color-gray;
+    }
+  }
+
+  ul {
+    background-color: $color-white;
+    border: 1px solid $color-gray;
+    border-radius: 5px;
+    padding: 2rem 2rem 2rem 4rem; // larger padding-left accounts for list-style
+  }
+
+  td ul {
+    border: none;
+  }
+}
